Create separate widget package for popup modules
WIDGETS FRNT-2 [IN PROGRESS]
