+++ BaCa2/BaCa2/db/ext_databases.py
@@ -9,16 +9,6 @@ DATABASES['owp_2023'] = {
 }
 
 
-DATABASES['ezk_2023'] = {
-    'ENGINE': 'django.db.backends.postgresql_psycopg2',
-    'NAME': 'ezk_2023_db',
-    'USER': 'baca2',
-    'PASSWORD': 'zaqwsxcde',
-    'HOST': 'localhost',
-    'PORT': ''
-}
-
-
 DATABASES['mj_2023'] = {
     'ENGINE': 'django.db.backends.postgresql_psycopg2',
     'NAME': 'mj_2023_db',
