+++ BaCa2/widgets/popups/forms.py
@@ -1,44 +1,12 @@
-from typing import Any, Dict, List
+from typing import List, Dict, Any
 
 from django.utils.translation import gettext_lazy as _
 
 from widgets.base import Widget
+from widgets.popups.base import PopupWidget
 
 
-class PopupWidget(Widget):
-    """
-    Base class for all popup widgets. Contains context fields common to all popups.
-
-    See Also:
-        - :class:`widgets.base.Widget`
-    """
-    def __init__(self,
-                 name: str,
-                 title: str,
-                 message: str,
-                 popup_class: str = "") -> None:
-        """
-        :param name: Name of the widget.
-        :type name: str
-        :param title: Title of the popup.
-        :type title: str
-        :param message: Message of the popup.
-        :type message: str
-        :param popup_class: CSS class applied to the popup.
-        :type popup_class: str
-        """
-        super().__init__(name=name, widget_class=popup_class)
-        self.title = title
-        self.message = message
-
-    def get_context(self) -> Dict[str, Any]:
-        return super().get_context() | {
-            'title': self.title,
-            'message': self.message,
-        }
-
-
-class FormConfirmationPopup(PopupWidget):
+class SubmitConfirmationPopup(PopupWidget):
     """
     A popup widget used to confirm a form submission. Can optionally display a summary of the form
     data.
